pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "soma"
version = "0.1.0"
description = "Sentient Observation & Memory Architecture — a self-constructing AI consciousness framework"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27,<1.0",
    "python-dotenv>=1.0",
    "lancedb>=0.13",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "streamlit>=1.38",
    "scikit-learn>=1.5",
    "scipy>=1.13",
    "bleak>=0.22",
    "fastapi>=0.115",
    "feedparser>=6.0",
    "requests>=2.31",
    "requests-oauthlib>=1.3",
    "pyjwt>=2.8",
    "pyedflib>=0.1.38",
    "uvicorn>=0.30",
    "mypy>=1.11",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
soma-research = "soma.autoresearcher.run:main"
soma-polar = "soma.proto_self.run:main"
soma-logger = "soma.proto_self.polar_logger:main"
soma-monitor = "soma.proto_self.monitor:main"
soma-baseline = "soma.proto_self.baseline_model:main"
soma-anomaly = "soma.proto_self.anomaly_detector:main"
soma-state = "soma.proto_self.state_classifier:main"
soma-recommend = "soma.proto_self.recommender:main"
soma-feedback = "soma.proto_self.feedback_logger:main"
soma-daily-research = "soma.proto_self.research_agent:main"
soma-review = "soma.proto_self.corpus_review:main"
soma-substack = "soma.proto_self.substack_agent:main"
soma-articles = "soma.proto_self.article_review:main"
soma-probe = "soma.proto_self.probe_interface:main"
soma-narrative = "soma.proto_self.narrative_builder:main"
soma-cpap-sync = "soma.proto_self.cpap.daily_sync:main"
soma-cpap-myair = "soma.proto_self.cpap.myair_client:main"
soma-cpap-edf = "soma.proto_self.cpap.edf_parser:main"
soma-profile = "soma.proto_self.soma_profile:main"

[tool.setuptools.packages.find]
include = ["soma*"]

[tool.mypy]
strict = true
packages = ["soma"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
